Move session and server into their own packages
